requirements.txt
streamlit>=1.31
pandas>=2.1
numpy>=1.26
openpyxl>=3.1
pyyaml>=6.0
python-dotenv>=1.0
